backend/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastapi-mail==1.5.0
fonttools>=4.50.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imutils==0.5.4
Jinja2==3.1.6
Levenshtein==0.27.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
numpy==1.26.4
asyncpg==0.29.0
SQLAlchemy==2.0.36
opencv-python<4.10
packaging==25.0
pillow==11.3.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.0
pyphen==0.17.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.14.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.37.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
